type discrepancy between man page and header file. libtiff version 3.8.2 defines the type as uint32 * and not u_long which triggers a compile warning on 64-bit platforms.

+++ src/simage_tiff.c
@@ -133,7 +133,7 @@ tiff_try_read_rgba(TIFF *in, int w, int h, int format,
     newbuffer = buffer;
   }
   if (!TIFFReadRGBAImage(in, w, h, 
-                         (unsigned long*) newbuffer, 1)) {
+                         (unsigned int*) newbuffer, 1)) {
     free(newbuffer);
     return ERR_READ;
   }
